+++ packages/cli/src/config/paths.ts
@@ -5,6 +5,7 @@ export default {
   appSrc: resolvePathInProject('./src'),
   appTsConfig: resolvePathInProject('./tsconfig.json'),
   appNodeModules: resolvePathInProject('node_modules'),
+  appPackageJson: resolvePathInProject('./package.json'),
 }
 
 const folders = {