+++ packages/example-ts-app/src/App.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import React, { useState, useEffect } from 'react'
 import cat from './cat.gif'
 import './App.scss'
 import Text from './Text'
@@ -10,6 +10,21 @@ interface AppProps {
 
 const App = ({ title }: AppProps) => {
   const [count, updateCount] = useState(0)
+  useEffect(() => {
+    // fullURL: 'https://www.zhihu.com/api/v4/mqtt/auth'
+    const testURL = '/api/v4/mqtt/auth'
+
+    fetch(testURL, {
+      method: 'POST',
+    })
+      .then(res => {
+        console.log('proxy works')
+      })
+      .catch(err => {
+        console.log('proxy not working')
+      })
+  }, [])
+
   return (
     <div className="app">
       <img src={cat} alt="cat" />